Showing that trace validation can fail after introducing consciously a thread bug in the implementation of KVS.
We comment a `synchronize` keyword on method `addOrReplace`, this can be lead to some transaction doing some work while another doing other work at the same time.

In this case, I don't think that it lead to a wrong implementation of KVS, because transaction `commit` method still `synchronized`. So, in principle, a transaction should never commit an update on a key that was committed first by another one. However, it doesn't respect the specification anymore: specification waiting that add or replace actions to be atomics.